+++ hortum_mobile/lib/views/complaint/complaint_page.dart
@@ -3,9 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:hortum_mobile/components/footer.dart';
 import 'package:hortum_mobile/components/spin.dart';
 import 'package:hortum_mobile/data/complaint_data_backend.dart';
-import 'package:hortum_mobile/services/codec_string.dart';
-import 'package:hortum_mobile/views/complaint/complaint_register_page.dart';
 import 'package:hortum_mobile/views/complaint/components/arrowback_button.dart';
+import 'package:hortum_mobile/views/complaint/components/create_complaint_button.dart';
 import 'components/list_complaints.dart';
 
 class ComplaintPage extends StatefulWidget {
@@ -21,8 +20,6 @@ class ComplaintPage extends StatefulWidget {
 class _ComplaintPageState extends State<ComplaintPage> {
   @override
   Widget build(BuildContext context) {
-    final TextEditingController name = new TextEditingController();
-    final TextEditingController description = new TextEditingController();
     Size size = MediaQuery.of(context).size;
     ComplaintDataAPI complaintDataAPI = new ComplaintDataAPI(widget.dio);
 
@@ -34,6 +31,10 @@ class _ComplaintPageState extends State<ComplaintPage> {
           body: Stack(
             children: [
               ArrowBackButton(),
+              CreateComplaintButton(
+                dio: widget.dio,
+                emailProductor: widget.emailProductor,
+              ),
               Container(
                 padding: EdgeInsets.only(
                   right: size.width * 0.1,
@@ -43,6 +44,7 @@ class _ComplaintPageState extends State<ComplaintPage> {
                   mainAxisAlignment: MainAxisAlignment.center,
                   children: [
                     Container(
+                      padding: EdgeInsets.only(bottom: size.height * 0.01),
                       child: Column(
                         children: [
                           Text(
@@ -54,48 +56,6 @@ class _ComplaintPageState extends State<ComplaintPage> {
                                 fontFamily: 'Comfortaa-Regular',
                                 letterSpacing: -0.33,
                                 fontWeight: FontWeight.w300),
-                          ),
-                          Container(
-                            alignment: Alignment.bottomRight,
-                            height: size.height * 0.06,
-                            padding: EdgeInsets.only(top: size.height * 0.005),
-                            child: MaterialButton(
-                              key: Key('createcomplaintButton'),
-                              onPressed: () {
-                                Navigator.push(
-                                  context,
-                                  MaterialPageRoute(
-                                    builder: (context) => ComplaintRegisterPage(
-                                      dio: widget.dio,
-                                      name: name,
-                                      description: description,
-                                      emailProductor:
-                                          decodeString(widget.emailProductor),
-                                    ),
-                                  ),
-                                );
-                              },
-                              child: Container(
-                                width: size.width * 0.3,
-                                height: size.height * 0.04,
-                                margin: EdgeInsets.only(
-                                    top: size.height / 100,
-                                    bottom: size.height / 100),
-                                decoration: BoxDecoration(
-                                  color: Color(0xffA7DDB7),
-                                  borderRadius:
-                                      BorderRadius.all(Radius.circular(20)),
-                                ),
-                                child: Align(
-                                  alignment: Alignment.center,
-                                  child: Text("Criar reclamação",
-                                      style: TextStyle(
-                                          color: Colors.black,
-                                          fontFamily: 'Roboto-Bold',
-                                          fontSize: 11)),
-                                ),
-                              ),
-                            ),
                           ),
                         ],
                       ),
